+++ gradle/check-licenses.gradle
@@ -130,7 +130,7 @@ downloadLicenses {
       'Eclipse Distribution License - v 1.0',
     ],
     (epl1): [
-      'Eclipse Public License - v 1.0'
+      'Eclipse Public License - v 1.0',
     ],
     (epl2): [
       'Eclipse Public License v2.0',
